:tada: "Add go.mod, LICENSE, dependabot.yml, and .gitignore files"
This commit adds the existing go.mod file which specifies the Go version (1.20) and the root module path. Also included is the LICENSE file, which lists rights and obligations according to the Apache License 2.0. The dependabot.yml file is added to enable dependency updates for Go modules with a daily schedule. A new .gitignore file is also attached which includes the standard rules for