:bug: Refactor jog.go to manage logging level init
Within the function New, the initialization and assignment code for logging level has been refactored to fix an out-of-index error. If 'level' slice is empty a crash. Now, default value 'LevelProd' assigned to 'll'. If 'level' slice is empty, 'll' is set to 'LevelDev'. This fix ensures the safety of index accessing on empty slices.